@@ -6,9 +6,6 @@ classifiers = [
     "Topic :: Software Development",
     "License :: OSI Approved :: Apache Software License",
     "Programming Language :: Python",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
 ]
 description = "Fast median finding"
 keywords = ["median", "finding", "quick", "select"]
@@ -16,40 +13,23 @@ license = "Apache Software License"
 name = "bfprt"
 readme = "README.md"
 repository = "https://github.com/gregorybchris/bfprt"
-version = "0.1.0"
+version = "0.2.0"
 
 [[tool.poetry.packages]]
 include = "bfprt"
 
 [tool.poetry.dependencies]
-click = "^8.1.3"
-python = ">=3.8.1,<4"
+python = ">=3.12"
 
 [tool.poetry.group.dev.dependencies]
-covcheck = "^0.4.2"
-flake8 = "^6.0.0"
-ipdb = "^0.13.9"
-mypy = "^1.0.1"
-pydocstyle = "^6.1.1"
-pylint = "^2.16.2"
-pyproject-flake8 = "^6.0.0"
-pytest = "^7.2.1"
-pytest-cov = "^4.0.0"
-semver = "^2.13.0"
-toml = "^0.10.2"
-yapf = "^0.32.0"
+mypy = "^1.10.0"
+pytest = "^8.2.0"
+ruff = "^0.5.0"
+semver = "^3.0.0"
 
 [build-system]
 build-backend = "poetry.core.masonry.api"
 requires = ["poetry-core>=1.0.0"]
-
-[tool.covcheck]
-branch = 100.0
-line = 100.0
-
-[tool.flake8]
-ignore = ""
-max-line-length = 120
 
 [tool.mypy]
 disallow_incomplete_defs = true
@@ -57,23 +37,51 @@ disallow_untyped_calls = true
 disallow_untyped_defs = true
 ignore_missing_imports = true
 
-[tool.pylint.basic]
-good-names = ["x, xs"]
-notes = ["FIXME"]
+[tool.ruff]
+line-length = 120
+src = ["bfprt"]
 
-[tool.pylint.format]
+[tool.ruff.lint]
+select = [
+    "A",
+    "ARG",
+    "B",
+    "BLE",
+    "C4",
+    "D",
+    "E",
+    "EM",
+    "ERA",
+    "F",
+    "FIX",
+    "G",
+    "I",
+    "ICN",
+    "ISC",
+    "LOG",
+    "N",
+    "PL",
+    "PT",
+    "PTH",
+    "PLR",
+    "RET",
+    "RUF",
+    "SIM",
+    "SLF",
+    "T20",
+    "TCH",
+    "TD",
+    "TID",
+    "W",
+]
+ignore = ["D203", "D213", "D100", "D104", "FIX002", "TD003"]
+
+
+[tool.ruff.lint.pycodestyle]
 max-line-length = 120
 
-[tool.pylint.master]
-extension-pkg-allow-list = ["pydantic"]
+[tool.ruff.lint.pydocstyle]
+convention = "google"
 
-[tool.pylint.messages_control]
-disable = ["missing-docstring"]
-
-[tool.pylint.reports]
-output-format = "colorized"
-
-[tool.yapf]
-based_on_style = "pep8"
-column_limit = 120
-indent_width = 4
+[tool.ruff.lint.extend-per-file-ignores]
+"**/tests/**/*.py" = ["D", "SLF"]
